call eval_do explicitly from fn_fn
There's no real reason to get the interned DO symbol and run an extra
loop of eval to see the special form. FN is already a special form. It
knows what kind of special it is. It can just call eval_do straight
up. However do_it in main.c is /not/ special enough to have that privilege.