added UNBOUND immediate value
With the advent of value slots on symbols, we should be able to bind a
symbol to the value NIL. Which means we cannot use NIL to track if
it's bound or not (which worked for the fvalue). This immediate value
cannot be produced by the read function, so a user could never
actually set a variable to this.