make global symbol_table
This serves a couple purposes. First, it cleans up signatures
everywhere. There's no real reason to pass the pointer to this down
the stack. It's global to the whole run of the program. Second, this
gives the GC easy access to the symbol table since the symbol table is
part of the root set.