don't pass Objection args as C args any more
With GC on the horizon, it's important that all lisp objects in use
actually be reachable from some set of root references. Consequently,
it becomes very important that we not rely on C runtime behavior such
as function arguments or return values.

Now that we have the PDA composed of current_expr and expr_stack, we
can eliminate the args parameter to all of our functions entirely.